Use Taskfile instead of Makefile
