+++ package.json
@@ -2,9 +2,9 @@
   "name": "voog-design-wellington",
   "description": "Basic boilerplate template for Voog",
   "scripts": {
-    "dev": "make dev",
-    "dev:kit": "make dev-kit",
-    "prod": "make prod"
+    "dev": "task dev",
+    "dev:kit": "task dev-kit",
+    "prod": "task prod"
   },
   "repository": {
     "type": "git",
@@ -12,11 +12,15 @@
   },
   "license": "ISC",
   "devDependencies": {
+    "@go-task/cli": "^3.30.1",
     "@rollup/plugin-node-resolve": "^15.2.1",
     "@rollup/plugin-terser": "^0.4.3",
     "chokidar-cli": "^3.0.0",
+    "cpy-cli": "^5.0.0",
+    "del-cli": "^5.1.0",
     "imagemin-cli": "^7.0.0",
     "rollup": "^3.29.2",
+    "rollup-plugin-concat": "^1.0.3",
     "sass": "^1.68.0"
   }
 }
